lib/requirements.txt
beautifulsoup4===4.6.0
googletrans===2.4.0
nltk===3.3
gensim===3.8.0
spacy===2.1.8
pandas===0.23.0
flask_sqlalchemy===2.3.2
pdfminer.six===20181108
pdf2image===1.8.0
pytesseract===0.3.0
anytree===2.6.0
dateparser===0.7.2
striprtf===0.0.5
textract===1.6.3
google-cloud-translate===2.0.0
